♻️[리팩토링] landing page, modal-component 연결
